refactor: Extract
