+++ app/charts/shared/render-value-labels.ts
@@ -1,73 +1,15 @@
 import { Selection } from "d3-selection";
-import { useMemo } from "react";
 
-import { ColumnsState } from "@/charts/column/columns-state";
-import { useChartState } from "@/charts/shared/chart-state";
 import {
   maybeTransition,
   RenderOptions,
 } from "@/charts/shared/rendering-utils";
-import { truthy } from "@/domain/types";
 
-type RenderValueLabelDatum = {
+export type RenderValueLabelDatum = {
   key: string;
   x: number;
   y: number;
   valueLabel: string;
-};
-
-export const useRenderValueLabelsData = () => {
-  const {
-    showValues,
-    renderEveryNthValue,
-    chartData,
-    getRenderingKey,
-    xScale,
-    getX,
-    yScale,
-    getY,
-    valueLabelFormatter,
-  } = useChartState() as ColumnsState;
-  const bandwidth = xScale.bandwidth();
-  const valueLabelsData: RenderValueLabelDatum[] = useMemo(() => {
-    if (!showValues) {
-      return [];
-    }
-
-    return chartData
-      .map((d, i) => {
-        if (i % renderEveryNthValue !== 0) {
-          return null;
-        }
-
-        const key = getRenderingKey(d);
-        const valueRaw = getY(d);
-        const xScaled = xScale(getX(d)) as number;
-        const value = valueRaw === null || isNaN(valueRaw) ? 0 : valueRaw;
-        const yRender = yScale(Math.max(value, 0));
-
-        return {
-          key,
-          x: xScaled + bandwidth / 2,
-          y: yRender,
-          valueLabel: valueLabelFormatter(value),
-        };
-      })
-      .filter(truthy);
-  }, [
-    showValues,
-    chartData,
-    renderEveryNthValue,
-    getRenderingKey,
-    getY,
-    xScale,
-    getX,
-    yScale,
-    bandwidth,
-    valueLabelFormatter,
-  ]);
-
-  return valueLabelsData;
 };
 
 export const renderValueLabels = (
